PSG: Emulate the PSG volumes "ideally" rather than matching some real systems' suboptimal output.
Most SMSs have bad audio output components, causing them to be unable to distinguish the upper volume levels. However, I think it is wrong to emulate this because some music will have been designed for the correct levels, and it seems wrong to emulate an analogue quality loss by default.