Add mapper `#46` MAPPER_SMS_Meta_Power_FFFF_HiCom
This mapper is a meta-multicart mapper for two 12-in-1 collections (one SMS-GG, the other SMS-MD) built from Korean Hi-Com 3-in-1 collections, and using cartridge power cycling to rotate to the next collection.

The 12-in-1 collections and their constituent parts are more thoroughly documented in https://github.com/ocornut/meka/pull/108

Right now this uses reset rather than power-cycle to trigger the outer multicart cycling. This is because Meka does not yet have hooks for mappers to respond reliably to power-cycling, nor does it have key bindings to allow easy use of power-cycling multicarts. For the SMS-GG collection running on the original Handy Gam*Boy hardware, there is no reset button so the use of emulator reset as a substitute for machine power-cycling seems somewhat reasonable. For the SMS-MD collection running on the original Super Aladdin Boy/Super Gam*Boy hardware, pressing the reset button should actually just return to the same multicart menu most recently seen, but this is not yet implemented.